Rewrite and update for Vagrant 1.4
There are a lot of changes in this patch. The makefile is now easier to
use, having only all, clean and per-.box targets that do what one would
think. They no longer automatically add anything to the host system so
there are less surprises.

The Vagrant.in file is gone, the way to use all the boxes is now
explained in the README.md. The default manifest and Vagrantfile are now
smarter, needing less configuration to build images if apt-cacher-ng is
installed on the local host.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>